HomeData可以上傳了

+++ 後端/HomeBackProject/HomeBackProject/Models/ADTypeData.cs
@@ -15,7 +15,6 @@ namespace HomeBackProject.Models
     using System.ComponentModel.DataAnnotations;
     using MetaDataHomeBackProject.Models;
     [MetadataType(typeof(MetaADTypeData))]
-
     public partial class ADTypeData
     {
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2214:DoNotCallOverridableMethodsInConstructors")]
@@ -26,7 +25,7 @@ namespace HomeBackProject.Models
             this.TerritoryData = new HashSet<TerritoryData>();
         }
     
-        public int ADID { get; set; }
+        public short ADID { get; set; }
         public string ADName { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
